First Release of Codes
## 02-02-2021

As the title states, this is the first release of the codes I've done in the month of classes from Jan 11th to Feb 2nd.

I've included two folders under the LCC_Classes_2020-24. One is where I'll store my school codes and the other is where I'll do a resources file in which I'll make a "compendium" of what I see during this course.

If anyone knows how to better word or how to describe the contents, please let me know.

Next update I'll add an English version of the Recursos file.